% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTsimrel-package.R
\docType{package}
\name{IRTsimrel-package}
\alias{IRTsimrel-package}
\alias{IRTsimrel}
\title{IRTsimrel: Reliability-Targeted Simulation for Item Response Data}
\description{
A framework for reliability-targeted simulation studies in Item Response
Theory (IRT). Researchers specify a target reliability level, and the
package calibrates item discriminations to achieve that target.
}
\details{
Main functions:
\itemize{
\item \code{\link{eqc_calibrate}}: Empirical Quadrature Calibration (recommended)
\item \code{\link{spc_calibrate}}: Stochastic Approximation Calibration
\item \code{\link{sim_latentG}}: Generate latent ability distributions
\item \code{\link{sim_item_params}}: Generate item parameters
\item \code{\link{simulate_response_data}}: Generate response data
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joonho112/IRTsimrel}
  \item Report bugs at \url{https://github.com/joonho112/IRTsimrel/issues}
}

}
\author{
JoonHo Lee
}
