% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_item_params.R
\name{print.summary.item_params}
\alias{print.summary.item_params}
\title{Print Method for summary.item_params Objects}
\usage{
\method{print}{summary.item_params}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{summary.item_params} object from \code{summary.item_params()}.}

\item{digits}{Integer. Number of decimal places for printing.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object, invisibly.
}
\description{
Print Method for summary.item_params Objects
}
