% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_latentG.R
\name{print.summary.latent_G}
\alias{print.summary.latent_G}
\title{Print Method for summary.latent_G Objects}
\usage{
\method{print}{summary.latent_G}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{summary.latent_G} object from \code{summary.latent_G()}.}

\item{digits}{Integer. Number of decimal places for printing.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object, invisibly.
}
\description{
Print Method for summary.latent_G Objects
}
