% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_item_params.R
\name{as.data.frame.item_params}
\alias{as.data.frame.item_params}
\title{Extract item parameters as data frame}
\usage{
\method{as.data.frame}{item_params}(x, ...)
}
\description{
Extract item parameters as data frame
}
