% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_item_params.R
\name{as.data.frame.item_params}
\alias{as.data.frame.item_params}
\title{Extract item parameters as data frame}
\usage{
\method{as.data.frame}{item_params}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An \code{item_params} object.}

\item{row.names}{NULL or a character vector giving the row names.}

\item{optional}{Logical. If TRUE, setting row names is optional.}

\item{...}{Additional arguments (ignored).}
}
\value{
A data frame of item parameters.
}
\description{
Extract item parameters as data frame
}
