% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_item_params.R
\name{plot.item_params}
\alias{plot.item_params}
\title{Plot method for item_params objects}
\usage{
\method{plot}{item_params}(x, type = c("scatter", "density", "both"), ...)
}
\description{
Plot method for item_params objects
}
