% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_latentG.R
\name{print.latent_G}
\alias{print.latent_G}
\title{Print Method for latent_G Objects}
\usage{
\method{print}{latent_G}(x, ...)
}
\arguments{
\item{x}{A \code{latent_G} object from \code{sim_latentG()}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The input object, invisibly.
}
\description{
Print Method for latent_G Objects
}
